requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
argon2-cffi=21.3.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
asttokens=2.0.5=pypi_0
attrs=21.4.0=pypi_0
backcall=0.2.0=pypi_0
black=21.12b0=pypi_0
blas=1.0=mkl
bleach=4.1.0=pypi_0
ca-certificates=2021.10.26=h06a4308_2
certifi=2021.10.8=py38h06a4308_2
cffi=1.15.0=pypi_0
charset-normalizer=2.0.10=pypi_0
click=8.0.3=pypi_0
cudatoolkit=11.3.1=h2bc3f7f_2
debugpy=1.5.1=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
entrypoints=0.3=pypi_0
executing=0.8.2=pypi_0
faiss-gpu=1.7.2=py3.8_h28a55e0_0_cuda11.3
filelock=3.4.2=pypi_0
flask=2.0.2=pypi_0
huggingface-hub=0.4.0=pypi_0
idna=3.3=pypi_0
importlib-resources=5.4.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.7.0=pypi_0
ipython=8.0.1=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=7.6.5=pypi_0
itsdangerous=2.0.1=pypi_0
jedi=0.18.1=pypi_0
jinja2=3.0.3=pypi_0
joblib=1.1.0=pypi_0
jsonschema=4.4.0=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=7.1.1=pypi_0
jupyter-console=6.4.0=pypi_0
jupyter-core=4.9.1=pypi_0
jupyterlab-pygments=0.1.2=pypi_0
jupyterlab-widgets=1.0.2=pypi_0
ld_impl_linux-64=2.35.1=h7274673_9
libfaiss=1.7.2=hfc2d529_0_cuda11.3
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgomp=9.3.0=h5101ec6_17
libstdcxx-ng=9.3.0=hd4cf53a_17
markupsafe=2.0.1=pypi_0
matplotlib-inline=0.1.3=pypi_0
mistune=0.8.4=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
mypy-extensions=0.4.3=pypi_0
nbclient=0.5.10=pypi_0
nbconvert=6.4.0=pypi_0
nbformat=5.1.3=pypi_0
ncurses=6.3=h7f8727e_2
nest-asyncio=1.5.4=pypi_0
nltk=3.6.7=pypi_0
notebook=6.4.7=pypi_0
numpy=1.22.1=pypi_0
numpy-base=1.21.2=py38h79a1101_0
openssl=1.1.1m=h7f8727e_0
packaging=21.3=pypi_0
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pypi_0
pathspec=0.9.0=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.0.0=pypi_0
pip=21.2.4=py38h06a4308_0
platformdirs=2.4.1=pypi_0
prometheus-client=0.12.0=pypi_0
prompt-toolkit=3.0.24=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.1=pypi_0
pycparser=2.21=pypi_0
pygments=2.11.2=pypi_0
pyparsing=3.0.6=pypi_0
pyrsistent=0.18.1=pypi_0
python=3.8.12=h12debd9_0
python-dateutil=2.8.2=pypi_0
pyyaml=6.0=pypi_0
pyzmq=22.3.0=pypi_0
qtconsole=5.2.2=pypi_0
qtpy=2.0.0=pypi_0
readline=8.1.2=h7f8727e_1
regex=2022.1.18=pypi_0
requests=2.27.1=pypi_0
sacremoses=0.0.47=pypi_0
scikit-learn=1.0.2=pypi_0
scipy=1.7.3=pypi_0
send2trash=1.8.0=pypi_0
sentence-transformers=2.1.0=pypi_0
sentencepiece=0.1.96=pypi_0
setuptools=58.0.4=py38h06a4308_0
six=1.16.0=pyhd3eb1b0_0
sqlite=3.37.0=hc218d9a_0
stack-data=0.1.4=pypi_0
terminado=0.12.1=pypi_0
testpath=0.5.0=pypi_0
threadpoolctl=3.0.0=pypi_0
tk=8.6.11=h1ccaba5_0
tokenizers=0.10.3=pypi_0
tomli=1.2.3=pypi_0
torch=1.10.1+cu113=pypi_0
torchaudio=0.10.1+cu113=pypi_0
torchvision=0.11.2+cu113=pypi_0
tornado=6.1=pypi_0
tqdm=4.62.3=pypi_0
traitlets=5.1.1=pypi_0
transformers=4.15.0=pypi_0
typing-extensions=4.0.1=pypi_0
urllib3=1.26.8=pypi_0
wcwidth=0.2.5=pypi_0
webencodings=0.5.1=pypi_0
werkzeug=2.0.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=3.5.2=pypi_0
xz=5.2.5=h7b6447c_0
zipp=3.7.0=pypi_0
zlib=1.2.11=h7f8727e_4
